+++ index.js
@@ -3,6 +3,7 @@ import { httpServer, app } from "./src/app";
 import errorHandler from "./src/middleware/error.middleware";
 import logger from "./src/logger";
 import { startCollection } from "./src/utils/metric";
+import { syncBound } from "./src/utils";
 import { currencyService } from "./src/services";
 
 setImmediate(() => {
@@ -15,6 +16,7 @@ setImmediate(() => {
             process.nextTick(async () => {
                 try {
                     await currencyService.updateCache();
+                    await syncBound();
                 } catch (e) {
                     logger.warning(e);
                 }