+++ src/controllers/bcb.controller.js
@@ -1,4 +1,4 @@
-import { bcbService } from "../services";
+import { bcbService, boundService } from "../services";
 import catchAsync from "../utils/catchAsync";
 
 export const selic = catchAsync(async (req, res) =>{
@@ -26,6 +26,16 @@ export const ifix = catchAsync(async (req, res) =>{
     res.json(data);
 });
 
+export const bdrx = catchAsync(async (req, res) =>{
+    const data = await bcbService.bdrx(req.query["type"]);
+    res.json(data);
+});
+
+export const sp500 = catchAsync(async (req, res) =>{
+    const data = await bcbService.sp500(req.query["type"]);
+    res.json(data);
+});
+
 export const ipca = catchAsync(async (req, res) =>{
     const data = await bcbService.ipca(req.query["type"]);
     res.json(data);
@@ -38,5 +48,15 @@ export const cdi = catchAsync(async (req, res) =>{
 
 export const news = catchAsync(async (req, res) =>{
     const data = await bcbService.news();
+    res.json(data);
+});
+
+export const boundList = catchAsync(async (req, res) =>{
+    const data = await boundService.findAll();
+    res.json(data);
+});
+
+export const bound = catchAsync(async (req, res) =>{
+    const data = await bcbService.bound(req.params.code, req.query["type"]);
     res.json(data);
 });