+++ src/validations/bcb.js
@@ -7,13 +7,15 @@ export const querySchema = joi.object({
     }),
 });
 
-export const queryIndicadorSchema = joi.object({
-    query: joi.object({
-        type: joi.string().valid("-1", "0", "1", "2", "3", "4")
-    }),
-});
-export const queryIndicador2Schema = joi.object({
-    query: joi.object({
-        type: joi.string().valid("0", "1", "2")
-    }),
-});
+/**
+ * 
+ * @param {Array<any>} valid 
+ * @returns 
+ */
+export const indicadorSchema = (valid)=>{
+    return joi.object({
+        query: joi.object({
+            type: joi.string().valid(...valid)
+        }),
+    });
+};